ACPI: Update /sys/firmware/acpi/interrupts/gpexx using new GPE forced disabling/enabling mechanism.
Now GPE can be forced enabling/disabling through acpi_force_gpe(), this
patch modifies /sys/firmware/acpi/interrupts/gpexx to use this new
facility. The "disable" command is implemented as forced disabling and the
"enable" command is implemented as forced flag resetting. "EN STS" is
returned to display the current hardware register status, along with "!"
flag to indicate the register bit unset and "*" flag to indicate the forced
disabling state.

The reason why we do not use forced enabling to implement "enable" command
is to avoid unexpected GPE storms. If the GPE is forced enabled, then
neither the OSPM driver nor the firmware provided method can disable it any
longer. If the GPE is level triggered, there will be cases for it to become
a GPE storm that hurts system performace, or lead to even worse effect in
single CPU environment.

Signed-off-by: Lv Zheng <lv.zheng@intel.com>