ACPI/EC: Add support to always poll events so that GPE can be always disabled and be enabled on demand.
While the EC driver is idle, we now still have 1 reference count enabled
by EVT_SCI. By converting EVT_SCI into pure polling mode, we can have GPE
disabled all the time while still can have all IBF/OBF handled in the
interrupt mode. This patch is a demonstration of this possiblity.

An interesting test is to cat gpe file twice after switching the EC driver
into the event polling mode:
  # echo Y > /sys/module/acpi/parameters/ec_poll_events
  # cat /sys/firmware/acpi/interrupts/gpe10
      1211     !EN     !STS   disabled
  # cat /sys/firmware/acpi/interrupts/gpe10
      1238     !EN     !STS   disabled
  # dmesg | tail
  [  370.697592] ACPI : EC: EC_SC(W) = 0x80
  [  370.697716] ACPI : EC: ===== IRQ (1) =====
  [  370.697733] ACPI : EC: EC_SC(R) = 0x00 SCI_EVT=0 BURST=0 CMD=0 IBF=0 OBF=0
  [  370.697738] ACPI : EC: EC_DATA(W) = 0x0c
  [  370.697828] ACPI : EC: ===== IRQ (1) =====
  [  370.697838] ACPI : EC: EC_SC(R) = 0x01 SCI_EVT=0 BURST=0 CMD=0 IBF=0 OBF=1
  [  370.697849] ACPI : EC: EC_DATA(R) = 0x00
  [  370.697920] ACPI : EC: ***** Command(RD_EC) stopped *****
  [  370.697937] ACPI : EC: 0: Decrease command
  [  370.698423] ACPI : EC: ##### Query(0x66) stopped #####
  # echo N > /sys/module/acpi/parameters/ec_poll_events
  # cat /sys/firmware/acpi/interrupts/gpe10
      1271      EN     !STS   enabled
We can see GPE is disabled by the driver and disabled in the register, but
we still can see interrupt count increased because all EC commands
(including the QR_EC issued by the event poller) are still processed in the
interrupt mode.
This patch can also pass system suspend/resume test even when the EC works
in the event polling mode.

If low power SoCs are capable of taking a disabled IRQ as an indication of
a very long period LTR, then the driver running in this mode can help the
SoCs to enter a system lower power state.

And we should note that recent low power SoCs will always split IRQs like
EVT_SCI as a seperate IRQ from the FIFO IRQs. This is because of a benefit
for the client platforms:
1. Since clients doesn't need to perform listening, as EC IBF/OBF, FIFO
   IRQs are normally can be disabled while still handled in the interrupt
   mode. This can and should be done by the software.
2. But exception IRQs, as EC EVT_SCI or USB bus reset, wifi beacon, need to
   be always enabled or they will have to be polled by the software using a
   timer. Instead of using a poller to waste CPU loads, it is better to
   still enable such IRQs for the software and offload its power saving
   work to the hardware. And because such IRQs are normally high latency
   tollerance, it won't prevent the SoCs from entering lower power state
   for longer period.

Note that this facility also can be used as a quirk mechanism for the buggy
EC firmware that cannot trigger GPE for unknown reasons.

Signed-off-by: Lv Zheng <lv.zheng@intel.com>