ACPICA: Utilities: Add deferred lock free contexts for deleting operand objects
When acpi_ut_delete_internal_obj() is invoked, reference count should be zero
which means this context is the only context remained referencing this
object.  Normally, we may just unlock namespace mutex (ACPI_MTX_NAMESPACE)
before invoking this function, but since this function is invoked from
different environment, doing this have side effects:
1. Other ACPI internal mutexes may be held orderly when this function is
   invoked. Without unlocking such locks, we still cannot achieve a lock
   free environment.
2. For some environment, it might not be safe to unlock the namespace
   mutex right in that context.
So we introduce a deferred unlocked environment for further cleanups.

There are two choices to convert this function into the lock free style:
1. Deferred work queue: this may not be performance friendly as in order to
   __delete__ the object, we __create__ a new OS object (work queue item)
   to achieve each deletion, which may lead to a failure when the memory
   resource is heavily used.
2. Garbage collection: the deleted objects are linked together and at each
   exit of such a deletion, a garbage cleaner function is invoked where
   this function is invoked in a lock free style.

This patch implements both choices:
1. OSL_GC_HANDLER: This is a deferred work item, it is used in case where
   the link point is not available for linking the deleted objects into the
   garbage list;
2. OSL_GC_THREAD: The linked deleted objects are destructed in a thread.
Note that in the slow path, we still need to __create__ a new OS object in
order to __delete__ the operand object. Lv Zheng.

Signed-off-by: Lv Zheng <lv.zheng@intel.com>
Cc: Lee Chun-Yi <jlee@suse.com>
Cc: Hanjun Guo <hanjun.guo@linaro.org>