ACPICA: ACPI 6.2: Add support for pin_group_config() resource
ACPICA commit 08b83591c0db751769d61fa889f4f50f575aeffb

pin_group_config() is analogous to pin_group_function() but instead of mode
(muxing), it is used to apply specific fine-grained configuration to a
set of referenced pins.

The format of this new resource is:

  pin_group_config (Shared/Exclusive, pin_config_type, pin_config_value,
                  resource_source, resource_source_index, resource_source_label,
                  resource_usage, descriptor_name, vendor_data)

The pin_config_type/pin_config_value are the same used by pin_config()
resource.

Here also the combination of resource_source and resource_source_label is
used to specify the pin_group() this resource refers to.

Link: https://github.com/acpica/acpica/commit/08b83591
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Signed-off-by: Lv Zheng <lv.zheng@intel.com>
Signed-off-by: Bob Moore <robert.moore@intel.com>