riscv: Add sfence.vma for mapping effection
When the last level of the MMU mapping (PTE) is changed, RISCV
requires significant barrier, otherwise a spurious fault may occur
or wrong mapping may be resulted when an OOO CPU lets instruction
that should be executed after the new mapping to be commited
before the PTE set operation. As instruction fetch is irrelative
to the memory write in OOO designs.

An fence.i or sfence.vma would be surfice to handle this case, so
this patch changes __set_fixmap() to allow sfence.vma to be applied
for both set and clear operations.

Signed-off-by: Lv Zheng <zhenglv@smart-core.cn>