irqchip/plic-sifive: Add DUOWEN platform quirks
It is tested that the current hardware implementation doesn't match the
original dual socket PLIC architecture. And thus cannot match the current
Linux spec driver:
1. If per-socket PLIC handles per-socket IRQ domain, the IRQ# and context
   ID differ from what the spec driver can handle.
2. The spec driver cannot handle both 1-256 and 258-513 IRQs to support
   IRQ redistribution to the remote socket with number of IRQs (riscv,ndev)
   limited to 256.

This patch is meant to fix the first issue to make the spec driver working
at least in the irq-socket-unshared mode. There is no easy way to fix the
second issue in the current PLIC driver as there is no mean to associate
multiple ENABLERs to a single context via a single probe.

In the dual socket mode, IRQ# is globally numbered from 1-256, 258-513.
The ENABLER only cares about local peripheral IRQs (e.x., 1-256 for PLIC on
socket0 and 258-513 for PLIC on socket1). And CLAIMR/COMPLETER only cares
about local socket context IDs.
This patch adds quirks to make PLIC driver to be adapted to such hardware
implementation.

Signed-off-by: Lv Zheng <zhenglv@smart-core.cn>
Reported-and-tested-by: Ian Jiang <jiangzhiying@smart-core.cn>